fix id error

+++ src/components/Product/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-
+import Swal from 'sweetalert2'
 // Redux
 import { useDispatch } from 'react-redux'
 import { deleteProductAction } from '../../actions/productsActions'
@@ -11,10 +11,21 @@ const Product = ({ name, price, id }) => {
 
   const confirmDeleteProduct = id => {
     // preguntar al usuario
-
-    // pasarlo al action
-    dispatch(deleteProductAction(id))
-    console.log(id)
+    Swal.fire({
+      title: 'Are you sure you want to delete the product?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if(result.value) {
+        // pasarlo al action
+        dispatch(deleteProductAction(id))
+        console.log(id)
+      }
+    })
   }
 
   return (
@@ -22,6 +33,7 @@ const Product = ({ name, price, id }) => {
       <td>{name}</td>
       <td>{price}</td>
       <td>
+        <td>{id}</td>
         <Link to={`/products/edit/${id}`}>Edit</Link>
         <button
           type='button'
